Move XcTest task implementation to the AbstractTestTask
