+++ subprojects/plugins/src/test/groovy/org/gradle/api/tasks/testing/TestTaskSpec.groovy
@@ -18,7 +18,7 @@ package org.gradle.api.tasks.testing
 
 import org.gradle.api.GradleException
 import org.gradle.api.internal.tasks.testing.*
-import org.gradle.api.internal.tasks.testing.detection.TestExecuter
+import org.gradle.api.internal.tasks.testing.TestExecuter
 import org.gradle.api.internal.tasks.testing.junit.report.TestReporter
 import org.gradle.internal.work.WorkerLeaseRegistry
 import org.gradle.test.fixtures.AbstractProjectBuilderSpec
