+++ subprojects/testing-base/src/main/java/org/gradle/api/internal/tasks/testing/TestExecuter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010 the original author or authors.
+ * Copyright 2017 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,10 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.api.internal.tasks.testing.detection;
-
-import org.gradle.api.internal.tasks.testing.TestExecutionSpec;
-import org.gradle.api.internal.tasks.testing.TestResultProcessor;
+package org.gradle.api.internal.tasks.testing;
 
 public interface TestExecuter<T extends TestExecutionSpec> {
     void execute(T testExecutionSpec, TestResultProcessor testResultProcessor);
