+++ subprojects/testing-base/src/main/java/org/gradle/api/internal/tasks/testing/junit/result/TestOutputAssociation.java
