+++ subprojects/testing-base/testing-base.gradle
@@ -20,6 +20,8 @@ dependencies {
     compile project(':core')
     compile project(':reporting')
     compile project(':platformBase')
+
+    compile libraries.kryo
 }
 
 useTestFixtures()
