+++ subprojects/testing-jvm/testing-jvm.gradle
@@ -23,7 +23,6 @@ dependencies {
     compile project(':testingBase')
 
     compile libraries.junit
-    compile libraries.kryo
     compile libraries.testng
 
     testCompile 'com.google.inject:guice:2.0@jar' // this is for TestNG
